Documentation: devel: add new project ideas section
This section will collect ideas to improve barebox and should serve as a
pool of ideas for people who want to enter the field of firmware development
but need some guidance what to work on.

These tasks can be adopted as part of programs like Google Summer of Code
or by motivated individuals outside such programs.

It's structured after coreboot's

  https://doc.coreboot.org/contributing/project_ideas.html

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220217082714.81749-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>