ARM: stm32mp: ddrctl: add STM32MP131 RAM size querying support
buswidth is read from HW. With nb_bytes == 2, there is a possibility
that we get a zero size out here, if driver and device are
mismatched, e.g. because barebox is booted in FIP with mismatched
external device tree.

As this runs very early before relocation, round up instead of crashing
to be a bit more on the safe side.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220223113846.3022227-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>