ARM: imx: Add imx8 support for SDRAM with two or more bank groups
Add bank group size to handle SDRAM with two or more bank groups
in one chip. The imx8mn DDR4 has one mem chip with ranks set to 1 and
the number of banks is 4 in 2 groups, total of 8 banks.

We need two add the DDRC_ADDRMAP8 and do a check DDRC_ADDRMAP8_BG_B0
and DDRC_ADDRMAP8_BG_B1 to get the number of bank groups in the chip.

Signed-off-by: Joacim Zetterling <joacim.zetterling@westermo.com>
Link: https://lore.barebox.org/20220225144751.4160843-3-joacim.zetterling@westermo.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>