ARM: arch: dts: qspi: Add QSPI config to the NXP IMX8MN-EVK board
Add QSPI support in the default config and DTS files for the
NXP IMX8MN-EVK board. To be able to use it You also need the
QSPI driver enabled.

Tested on the NXP IMX8MN-EVK board as well as on an IMX8MN
custom bord supporting a QSPI flash. Tested with the memory
dump command i.e. 'md -s /dev/m25p0'.

Signed-off-by: Joacim Zetterling <joacim.zetterling@westermo.com>
Link: https://lore.barebox.org/20220301132308.343496-1-joacim.zetterling@westermo.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>