partitions: efi: print GPT offset in debug message
This debug message is printed for the alternative GPT header as well,
which may not exist at that location, because the medium has a different
length than what's in the GPT header. Make it easier to differentiate,
whether the message refers to the primary or alternate GPT header by
printing the LBA the header was expected at.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Link: https://lore.barebox.org/20220319072427.1310855-1-ahmad@a3f.at
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>