ARM: fix GCC 11.x build failures for ARMv7
Building barebox for ARMv7 with GCC 11.x results in errors similar to
the following:

  CC      common/state/state_variables.o
{standard input}: Assembler messages:
{standard input}:1535: Error: selected processor does not support `rev r3,r3' in Thumb mode
{standard input}:1576: Error: selected processor does not support `rev r3,r3' in Thumb mode

Or:

  CC      common/state/state_variables.o
{standard input}: Assembler messages:
{standard input}:1405: Error: selected processor does not support `rev r3,r3' in ARM mode
{standard input}:1453: Error: selected processor does not support `rev r3,r3' in ARM mode

The problem is that the compiler option `-march=armv7-a` is not being
chosen by "arch/arm/Makefile", but rather the fallback options
`-march=armv5t -Wa,-march=armv7-a` are being chosen.

Appending `-msoft-float` to `KBUILD_CPPFLAGS` earlier in
"arch/arm/Makefile" before the tests for supported `-march` options
seems to fix the problem.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Link: https://lore.barebox.org/20220408170154.114526-1-abbotti@mev.co.uk
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>