ARM: rpi: reinstate support for multiple variants with same DT
Raspberry Pi support contains custom board code to allow the same DT to
be used across different board model for a given Raspberry Pi
generation. For this to work, barebox may not deduce the model specific
init from the device tree compatible, but instead detect it
automatically. This was broken in the commit changing the board code to
the driver model, so reinstate it.

Fixes: c062cd5cf47d ("ARM: rpi: validate devicetree compatible instead of changing model name")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220530162016.1052586-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>