commands: ls: explicitly check for directories with S_ISDIR
S_ISDIR(mode) is implemented as (mode & S_IFMT) == S_IFDIR, which
accounts for file modes setting multiple bits. So far, this was not the
case, but upcoming S_IFBLK equals (S_IFCHR | S_IFDIR), which would fail
the existing checks, so prepare for that by fixing them.

No functional change. There are no other instances of this elsewhere in
the code base.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220602195916.9061-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>