arm: imx8mq: re-enable DDRC for Barebox
With Linux 5.18 the DDRC is disabled for most boards in the upstream
DT, as it isn't used. Barebox however needs the DDRC node to be enabled
as it has a driver to read the memory size from the controller and ends
up with no available RAM without this driver being present.

Signed-off-by: Lucas Stach <dev@lynxeye.de>
Link: https://lore.barebox.org/20220609214330.1443541-1-dev@lynxeye.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>