net: designware: rockchip: future-proof driver for more SoCs
Linux driver has a struct rk_gmac_ops::regs_valid member to mark
whether regs is populated or not. Population happens via a GCC
extension that allows static initialization of flexible array
members. Replace that with a pointer that is checked, so future
SoC support just initializes it to NULL by default and avoid
the OOB access.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220615080305.954879-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>