mci: imx-esdhc-common: Don't clear unhandled status bits
A DMA cmd + data transfer can finish or stop (i.e. on a block gap) before
the status register is cleared. In that case we'll lose track of state
causing sdhci_transfer_data_dma() to loop forever waiting for status bits
that are already cleared.

Clearing SDHCI_INT_CMD_COMPLETE should suffice here. Since it's not
evaluated a second time, clearing it at the start of the function is
sufficient so we can just remove the erroneous status write.

Signed-off-by: Robin van der Gracht <robin@protonic.nl>
Link: https://lore.barebox.org/20220621101529.554322-1-robin@protonic.nl
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>