ARM: cpu: start: don't panic when only initial memory known
Boards that get their memory from device tree are prone to breakage if
upstream decides to remove the /memory node or renders it unusable[0]
or to disable the DRAM controller device node[1] used to dynamically
compute the size.  While this needs to be fixed properly on a
case-by-case basis, soften the blow by using the memory size specified
by low level code as a fallback instead of just throwing the towel.

[0]: aa68c6bd3615 ("arm: rk3288-phycore-som: Fix memory node")
[1]: 0529149b2756 ("arm: imx8mq: re-enable DDRC for Barebox")

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220622175111.2657673-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>