ARM: i.MX8M: refactor to prepare i.MX8MN LPDDR4 support
Device config has no effect on i.MX7 and (i.MX8M && LPDDR4).
In these configuration, it's assumed to be 4, which isn't correct for
the i.MX8MN, which has a 16-bit bus. Instead of complicating the
function further, allow it to assume device config is always correct and
have the callers worry about fixing it up as necessary.

No functional change.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220623130717.1447999-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>