ARM: i.MX8MP: adapt atf bl31 base address
With i.MX8MP A1 silicon the OCRAM space is extended to 576KB whereas
the pre-release silicon has only 512KB. So the upstream TF-A adapted the
base address to move BL31 to the last 128KB of the OCRAM.

The adaption was a bit messy because each version changed it:
 - v2.4 base addr = 0x960000 (original)
 - v2.5 base addr = 0x970000 (adapted to new silicon revision)
 - v2.6 base addr = 0x960000 (changed back by accident)
 - v2.7 base addr = 0x970000 (accident fixed)

With v2.7 it is correctly set for A1 silicon onwards and we strongly
recommend to use this version or newer.

This commit also adapts the documentation for the i.MX8MP-EVK to
reference the upstream TF-A and to point out our favorite version.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20220704165256.589530-1-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>