usb: chipidea: imx: call enable_clk after all resources are requested
Currently we enable the clock immediately after requesting it and leave
it on the whole time. Afterwards if the phy request is failing we leave
the usb-controller <-> usb-phy connection in a partly initialized state.
At least on i.MX8M SoCs this can cause strange system hangs during boot.

The (more) correct way would be to have the whole power-domain framework
and the blk-ctrl driver support within barebox. So we can leave the
power-domain in a known good state for linux. Since this is not the case
we can move the clk_enable() call so it gets called after we know that
all ressources are available. So the probability to leave the system in
a partly initialized state is lesser.

Drop the !IS_ERR() since NULL and errors are handled in clk_enable() as
well.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Tested-by: Johannes Zink <j.zink@pengutronix.de> # i.MX7
Link: https://lore.barebox.org/20220706142105.2266956-4-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>