ARM: i.MX8MN: pass along correct DT depending on variant
We support two different i.MX8MN variants of the i.MX8MN-EVK:
One with LPDDR4 and another with DDR4, each with a different PMIC.
The PMICs are at different i2c addresses, which allows us to
differentiate between the variants in PBL, but in barebox proper, we use
the same DT for both. Fix that, so we run with a device tree that
reflects the actual hardware. This allows us to apply the esdctl
quirk selectively too.

Tested on the DDR4 variant (Marco)

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
m.felsch@pengutronix.de: added deep-probe
m.felsch@pengutronix.de: rebased on next
Tested-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20220706142105.2266956-5-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>