regulator: handle regulator_get_voltage(NULL) gracefully
Unlike recent versions of Linux, barebox represents the dummy regulator
as a NULL pointer and regulator API is supposed to anticipate operation
on a NULL pointer. When regulator_get_voltage was ported from Linux,
this was not taken into consideration, which leads to
regulator_get_voltage(NULL) crashing with a NULL pointer dereference.

Fix this by returning -EINVAL for dummy regulators. This aligns us with
how Linux would behave in this situation.

Fixes: 6ee83ce08b24 ("regulator: add regulator_get_voltage() to API")
Reported-by: Johannes Zink <j.zink@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20220718114824.2632364-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>